ai-testing-agent/backend/requirements.txt
Flask==3.0.0
flask-cors>=4.0.0
openai>=1.12.0
httpx<0.28
python-dotenv
flask
requests
flask-migrate>=4.0.0
flask-sqlalchemy>=3.0.0
psycopg[binary]>=3.1.0
bcrypt>=4.0.0
PyJWT>=2.8.0
cryptography>=3.0.0

